Extract stubbing from DataRequest into stub/non-stub modules
