+++ lib/do_api_scripting/api/all_droplets/data_request.rb
@@ -2,6 +2,9 @@
 
 require 'excon'
 
+require_relative './non_stubs'
+require_relative './stubs'
+
 # Code to support scripting the DigitalOcean API, e.g., for use with Ansible.
 module DoApiScripting
   module API
@@ -9,77 +12,20 @@ module DoApiScripting
     class AllDroplets
       # Class encapsulating sending request to/receiving response from DO API.
       class DataRequest
-        # Stubbing data shouldn't be part of the core logic, should it? :P
-        module Stubs
-          def self.request_params(headers)
-            { mock: true, headers: headers }
-          end
-
-          def self.on
-            Excon.stub({}, status: 200, body: JSON.dump(DUMMY_DATA))
-          end
-
-          def self.off
-            Excon.stubs.clear
-          end
-
-          DUMMY_DATA = {
-            meta: { total: 3 },
-            links: {},
-            droplets: [
-              {
-                created_at: '2017-06-21T08:20:42Z',
-                id: '52434906',
-                name: 'suirdemo1-test',
-                networks: {
-                  v4: [
-                    { type: 'private', ip_address: '10.130.10.113' },
-                    { type: 'public', ip_address: '128.199.105.180' }
-                  ],
-                  v6: []
-                },
-                region: { name: 'Singapore 1', slug: 'sgp1' },
-                size_slug: '512mb',
-                status: 'active'
-              },
-              {
-                created_at: '2017-06-22T10:07:34Z',
-                id: '52569259',
-                name: 'suirdemo2',
-                networks: {
-                  v4: [
-                    { type: 'private', ip_address: '10.130.19.125' },
-                    { type: 'public', ip_address: '128.199.73.100' }
-                  ],
-                  v6: []
-                },
-                region: { name: 'Singapore 1', slug: 'sgp1' },
-                size_slug: '512mb',
-                status: 'active'
-              }
-            ]
-          }.freeze
-          private_constant :DUMMY_DATA
+        def self.get(auth_header: :from_env, request_module: Stubs)
+          new(auth_header, request_module).get
         end
 
-        def self.get(auth_header = :from_env)
-          new(auth_header).get
-        end
-
-        # Reek points out that this is (temporarily) a :reek:UtilityFunction.
         def get
-          Stubs.on
-          url = 'https://api.digitalocean.com/v2/droplets'
-          resp = Excon.get(url, stubs.request_params(headers))
-          Stubs.off
-          resp
+          stubs.request(headers: headers, url: URL)
         end
 
         protected
 
-        def initialize(auth_header)
-          @auth_header = auth_header
-          if auth_header == :default
+        def initialize(auth_header_in, stub_module)
+          @stubs = stub_module
+          @auth_header = auth_header_in
+          if auth_header_in == :from_env
             @auth_header = "Bearer #{ENV['DO_API_TOKEN']}"
           end
           self
@@ -87,7 +33,10 @@ module DoApiScripting
 
         private
 
-        attr_reader :auth_header
+        attr_reader :auth_header, :stubs
+
+        URL = 'https://api.digitalocean.com/v2/droplets'
+        private_constant :URL
 
         def headers
           {
